==> notebooks/1_scrape.ipynb <==
# Scraping

This notebook uses routines from [```scrape.py```](../util/scrape.py) to scrape fanfiction from [Archive of our Own](http://archiveofourown.org) (AO3). We've chosen to focus on <i>Avatar: The Last Airbender</i> (which we'll call Avatar) fanfiction for this project.

import sys
sys.path.append("..")

from util.scrape import get_search_url, get_works_info, scrape_fic_list
import numpy as np

%load_ext autoreload
%autoreload 2

We define our search query: completed Avatar fanfiction ("fics") with at least 400 kudos. We restrict ourselves to English-language works.

avatar_search = get_search_url("Avatar: The Last Airbender", min_kudos=400, complete=True, english_only=True)

avatar_search

Let's scrape the search page to collect metadata for fics we would like to include in our corpus. Here we restrict our collection to works that are not in a series (some authors post dozens of chapters as individual fics, which biases our data towards those few authors) and are longer than 1000 words.

works_info = get_works_info(avatar_search, 2000, word_range=(1000,0), exclude_series=True)

The above was run 11/15/20 at 7:42 AM, but wasn't saved and a rerun would be different. Let's see an example of the format.

example = get_works_info(avatar_search, 20, exclude_series=True)

example[0]

We pass an array of similar metadata to ```scrape_fic_list```, which will use ```work_id``` to find the fanfiction's URL and scrape its text.

fic_df = scrape_fic_list(works_info, print_every=30)

Here's what the first few entries look like.

fic_df.head(3)

Save our dataframe to a file.

#fic_df.to_pickle("../data/avatar_fics_scraped.pickle")

Potential future work: don't put too many works from a single author in the corpus (especially if they're very short).
