==> notebooks/2_preprocess.ipynb <==
# Preprocessing

We preprocess the text we scraped in [```1_scrape.py```](1_scrape.py), converting it into a format usable for our NLP modeling.

import pandas as pd

import sys
sys.path.append("..")
from util.preprocess import process_texts, find_ngrams

%load_ext autoreload
%autoreload 2


Let's load our data.

fic_df = pd.read_pickle("../data/avatar_fics_scraped.pickle")

Our preprocessing logic is contained in [```preprocessing.py```](../util/preprocessing.py), which cleans, tokenizes, lemmatizes, removes stopwords, and merges specified n-grams. Most of the process is automatic, but we do need to specify the n-grams we would like merged.

### Finding n-grams

Let's use ```find_ngrams``` to find the most common bigrams and trigrams in the corpus. We can examine those for ideas on what to merge.

# Exclude "ten" as a stopword because "Lu Ten" is an imporant character in the series.
bigrams, trigrams = find_ngrams(fic_df["text"], stop_exceptions = ["ten"])

bigrams.most_common()[0:400]

trigrams.most_common()[:100]

Based on the above counters and on domain knowledge, we decide on bi- and trigrams. We also fix certain lemmatization errors. 

avatar_bigrams = ["fire lord", "fire nation", "lu ten", "water tribe", "ty lee", "lu ten", "earth kingdom", "water tribe", "agni kai", "blue spirit", "air temple", "boiling rock", "chit sang", "ember island", "dai li", "pai sho", "jasmine dragon", "long feng", "north pole", "south pole", "spirit world", "white lotus", "gran gran", "si wong", "white lotus", "air nomad"]
avatar_trigrams = ["ba sing se" ]

elements = ["", "water", "earth", "fire", "air"]

# unigram exceptions
u_exc_1 = {f"{element}benders":f"{element}bender" for element in elements}
u_exc_2 = {f"{element}bends":f"{element}bend" for element in elements}
u_exc_3 = {f"{element}bending":f"{element}bend" for element in elements}
u_exc_4 = {f"{element}bended": f"{element}bend" for element in elements}
u_exc_5 = {"soulmates": "soulmate", "firelord": "fire lord", "firelords": "fire lord", "grangran": "gran gran"}
u_exceptions = {**u_exc_1, **u_exc_2, **u_exc_3, **u_exc_4, **u_exc_5}

# bigram exceptions
b_exceptions = {"soul mate": "soulmate", "soul mates": "soulmate", "gon na": "gonna", "wan na": "wanna", "agni kais": "agni kai"}

avatar_exceptions = {**u_exceptions, **b_exceptions}

avatar_ngrams = {"bigrams": avatar_bigrams, "trigrams": avatar_trigrams, "exceptions": avatar_exceptions}


A very specific case is the bigram "Lu Ten," a character in Avatar. "ten" is included in our default stopword list, so we remove it from the stopword list. Then, after merging bigrams, we remove all instances of "ten" that weren't merged into a bigram.

This functionality can be extended to other stopwords that appear in important ngrams.

avatar_stop_exceptions = ["ten"]
avatar_stop_after = ["ten"]

### Example Process

Let's look at an example of how the text is processed.

fic_df.iloc[5]

text_example = fic_df.iloc[5]["text"][:505]
text_example

processed_example = process_texts([text_start], ngrams=avatar_ngrams, stop_exceptions=avatar_stop_exceptions, stop_after = avatar_stop_after)
print(processed_example)

### Final Processing

We run our text through the preprocessing sequence, add it to our dataframe, and pickle the result.

processed_texts = process_texts(fic_df["text"], ngrams=avatar_ngrams, stop_exceptions=avatar_stop_exceptions, stop_after = avatar_stop_after)
fic_df["processed"] = processed_texts

fic_df.to_pickle("../data/avatar_fics_processed.pickle")
