==> notebooks/3_lda.ipynb <==
# Topic Modeling via Latent Dirichlet Allocation (LDA)

We model topics for our dataset using LDA, selecting hyperparameters for the model using the ```c_v``` coherence metric.

See [this paper](http://svn.aksw.org/papers/2015/WSDM_Topic_Evaluation/public.pdf) for a definition of the various coherence metrics implemented in Gensim's [```CoherenceModel```](https://radimrehurek.com/gensim/models/coherencemodel.html).

import pandas as pd
import sys
sys.path.append("..")

from util.model import LDATuner
from itertools import product
import pickle

%load_ext autoreload
%autoreload 2

df = pd.read_pickle("../data/avatar_fics_processed.pickle")
docs = df["processed"]

We use our ```LDATuner``` class from [```util/model.py```](../util/model.py). The parameters we tune are ```num_topics```, the number of topics in the model, ```min_thresh```, the minimum number of documents a word must appear in to be included in the dictionary, and ```max_thresh```, the maximum percentage of documents a word can appear in before being excluded from the dictionary.

Some external testing was conducted for ```alpha``` and ```beta``` and we concluded that ```asymmetric``` and ```symmetric``` were the superior options.

tune = LDATuner(docs, verbose=False)
keywords=["num_topics", "min_thresh", "max_thresh", "alpha", "beta"]

Let's fix a minimum and maximum threshold and see which number of topics seems best. 

It won't necessarily be best globally as we have fixed our thresholds, but we use it here as an approximation and an idea of where to conduct a search with greater granularity.

num_topics = [8, 9, 10, 11, 12, 13, 14]
min_thresh = [25]
max_thresh = [0.3]

alpha=['asymmetric']
beta=['symmetric']

count = 0

iterations = 400
passes = 20
 
keywords=["num_topics", "min_thresh", "max_thresh", "alpha", "beta"]
param_grid = product(num_topics, min_thresh, max_thresh, alpha, beta)

model_scores = tune.tune_hyper(keywords, param_grid)

num_topics = [15, 16, 17, 18, 19, 20]
min_thresh = [25]
max_thresh = [0.3]

alpha=['asymmetric']
beta=['symmetric']

count = 0

iterations = 400
passes = 20
 
keywords=["num_topics", "min_thresh", "max_thresh", "alpha", "beta"]
param_grid = product(num_topics, min_thresh, max_thresh, alpha, beta)

model_scores = tune.tune_hyper(keywords, param_grid)

We get the best results with 11 and 12 topics. Let's do some more fine-grained searching in the 11- and 12-topic range, this time focusing our efforts on tuning the dictionary inclusion thresholds.

num_topics = [11, 12]
min_thresh = [20, 30, 40, 50]
max_thresh = [0.25, 0.3, 0.35]
alpha=['asymmetric']
beta=['symmetric']

iterations = 400
passes = 20
 
keywords=["num_topics", "min_thresh", "max_thresh", "alpha", "beta"]
param_grid = product(num_topics, min_thresh, max_thresh, alpha, beta)

model_scores = tune.tune_hyper(keywords, param_grid)

#pickle.dump(model_scores, open("../models/avatar_model_scores.pickle", "wb" ) )
model_scores = pickle.load(open("../models/avatar_model_scores.pickle", "rb"))

We've found the several good solutions (by ``c_v`` coherence) among the various parameter combinations we evaluated. Let's train a few of the best, then evaluate them by comparing the topics to our domain knowledge.

We don't automatically select the best model by coherence because the metric isn't necessarily the most indicative of what a human would judge to be a good topic model. In fact, the original paper evaluates coherence metrics by their correlation with human ratings; ```c_v``` was the highest at 0.731.

model_info_1 = tune.train_model_params((11, 50, 0.25, 'asymmetric', 'symmetric'), keywords)
model_info_2 = tune.train_model_params((11, 50, 0.3, 'asymmetric', 'symmetric'), keywords)
model_info_3 = tune.train_model_params((12, 50, 0.3, 'asymmetric', 'symmetric'), keywords)
model_info_4 = tune.train_model_params((12, 50, 0.35, 'asymmetric', 'symmetric'), keywords)

What do our topics look like?

We print the topics for each of the top 4 models, then evaluate them by an eye test. We omit the in-depth reasoning as that would involve detailed domain knowledge.

mod_1 = model_info_1["model"]
for i in range(mod_1.num_topics):
    print(mod_1.show_topic(i, topn=20), "\n")

Our best model actaully looks fairly bad; several areas seem mixed together in certain topics based on domain knowledge.

mod_2 = model_info_2["model"]
for i in range(mod_2.num_topics):
    print(mod_2.show_topic(i, topn=20), "\n")

Model 2 seems to have fairly well-grouped topics!

mod_3 = model_info_3["model"]
for i in range(mod_3.num_topics):
    print(mod_3.show_topic(i, topn=20), "\n")

mod_4 = model_info_4["model"]
for i in range(mod_4.num_topics):
    print(mod_4.show_topic(i, topn=20), "\n")

Models 3 and 4 are slightly worse than model 2, subjectively. 

We choose model 2 as our final model. We'll analyze the model and cluster the data points in [```4_cluster.ipynb```](4_cluster.ipynb).

model_info = model_info_2
pickle.dump(model_info, open("../models/avatar_model.pickle", "wb" ) )
